Implement MR comments
