+++ newton/viewer.py
@@ -2,6 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # Import all viewer classes (they handle missing dependencies at instantiation time)
+from ._src import viewer as _viewer
 from ._src.viewer import (
     Layer,
     ViewerBase,
@@ -13,6 +14,9 @@ from ._src.viewer import (
     ViewerUSD,
     ViewerViser,
 )
+
+_BUILTIN_VIEWER_NAMES = _viewer._BUILTIN_VIEWER_NAMES
+_get_viewer_entry_points = _viewer._get_viewer_entry_points
 
 __all__ = [
     "Layer",
